Tendar_project/src/requirement.txt.txt
async-generator==1.10
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
botocore==1.29.126
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
exceptiongroup==1.1.1
filelock==3.12.0
fonttools==4.39.3
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
parsel==1.8.1
Pillow==9.5.0
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
s3transfer==0.6.0
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uuid==1.30
w3lib==2.1.1
wsproto==1.2.0
zope.interface==6.0